requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
blinker==1.4
certifi==2018.11.29
chardet==3.0.4
Click==7.0
coverage==4.5.2
coveralls==1.5.1
docopt==0.6.2
Flask==1.0.2
Flask-Mail==0.9.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
more-itertools==5.0.0
pathlib2==2.3.3
pluggy==0.8.1
psycopg2==2.7.6.1
py==1.7.0
PyJWT==1.7.1
pytest==4.1.1
pytest-cov==2.6.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
